server/file_service/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Cors
Werkzeug
python-dotenv
gunicorn
psycopg2-binary